# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "DailyChallenge": {
        "properties": {
          "answer_tag": {
            "title": "Answer Tag",
            "type": "string"
          },
          "challenge_id": {
            "title": "Challenge Id",
            "type": "string"
          },
          "clues": {
            "$ref": "#/components/schemas/DailyChallengeClues"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/DailyChallengeImage"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/DailyChallengeOption"
            },
            "title": "Options",
            "type": "array"
          },
          "total_candidates": {
            "title": "Total Candidates",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "challenge_id",
          "image",
          "answer_tag",
          "total_candidates"
        ],
        "title": "DailyChallenge",
        "type": "object"
      },
      "DailyChallengeClues": {
        "properties": {
          "artists": {
            "items": {
              "type": "string"
            },
            "title": "Artists",
            "type": "array"
          },
          "copyrights": {
            "items": {
              "type": "string"
            },
            "title": "Copyrights",
            "type": "array"
          },
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "general": {
            "items": {
              "type": "string"
            },
            "title": "General",
            "type": "array"
          },
          "meta": {
            "items": {
              "type": "string"
            },
            "title": "Meta",
            "type": "array"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "title": "DailyChallengeClues",
        "type": "object"
      },
      "DailyChallengeImage": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "danbooru_post_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Danbooru Post Id"
          },
          "ext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext"
          },
          "file_id": {
            "title": "File Id",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "thumbnail_token": {
            "title": "Thumbnail Token",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "file_id",
          "thumbnail_token",
          "filename",
          "folder",
          "ext",
          "width",
          "height",
          "score",
          "rating",
          "created_at",
          "danbooru_post_id"
        ],
        "title": "DailyChallengeImage",
        "type": "object"
      },
      "DailyChallengeOption": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "count"
        ],
        "title": "DailyChallengeOption",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HomeImageRail": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HomeImageRailItem"
            },
            "title": "Items",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "title": "HomeImageRail",
        "type": "object"
      },
      "HomeImageRailItem": {
        "properties": {
          "ext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext"
          },
          "file_id": {
            "title": "File Id",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_favorite": {
            "default": false,
            "title": "Is Favorite",
            "type": "boolean"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "tag_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Category"
          },
          "tag_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Name"
          },
          "thumbnail_token": {
            "title": "Thumbnail Token",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "file_id",
          "thumbnail_token",
          "filename",
          "folder",
          "ext",
          "width",
          "height",
          "score",
          "rating"
        ],
        "title": "HomeImageRailItem",
        "type": "object"
      },
      "HomeImageRails": {
        "properties": {
          "rails": {
            "items": {
              "$ref": "#/components/schemas/HomeImageRail"
            },
            "title": "Rails",
            "type": "array"
          }
        },
        "title": "HomeImageRails",
        "type": "object"
      },
      "HomeTagInfo": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "cover_candidates": {
            "items": {
              "$ref": "#/components/schemas/HomeCoverCandidate"
            },
            "title": "Cover Candidates",
            "type": "array"
          },
          "cover_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover File Id"
          },
          "cover_post_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Post Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "thumbnail_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Token"
          }
        },
        "required": [
          "name",
          "category",
          "count"
        ],
        "title": "HomeTagInfo",
        "type": "object"
      },
      "HomeTags": {
        "properties": {
          "featured": {
            "items": {
              "$ref": "#/components/schemas/HomeTagInfo"
            },
            "title": "Featured",
            "type": "array"
          },
          "groups": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/HomeTagInfo"
              },
              "type": "array"
            },
            "title": "Groups",
            "type": "object"
          }
        },
        "title": "HomeTags",
        "type": "object"
      },
...
      "PopularityPeriod": {
        "properties": {
          "average_score": {
            "title": "Average Score",
            "type": "number"
          },
          "best_score": {
            "title": "Best Score",
            "type": "integer"
          },
          "end_date": {
            "title": "End Date",
            "type": "string"
          },
          "image_count": {
            "title": "Image Count",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "popularity": {
            "title": "Popularity",
            "type": "integer"
          },
          "start_date": {
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "period",
          "label",
          "start_date",
          "end_date",
          "image_count",
          "popularity",
          "average_score",
          "best_score"
        ],
        "title": "PopularityPeriod",
        "type": "object"
      },
...
      "TagInfo": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "favorite_added_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorite Added At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pinned_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned At"
          }
        },
        "required": [
          "name",
          "category",
          "count"
        ],
        "title": "TagInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Keivotos - Danbooru",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/challenges/characters/suggest": {
      "get": {
        "operationId": "challenge_character_suggestions_api_challenges_characters_suggest_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:[gsqeu](?:,[gsqeu])*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 30,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagInfo"
                  },
                  "title": "Response Challenge Character Suggestions Api Challenges Characters Suggest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Challenge Character Suggestions"
      }
    },
    "/api/challenges/daily": {
      "get": {
        "operationId": "daily_challenge_api_challenges_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:[gsqeu](?:,[gsqeu])*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyChallenge"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily Challenge"
      }
    },
...
    "/api/home/image-rails": {
      "get": {
        "operationId": "home_image_rails_api_home_image_rails_get",
        "parameters": [
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:[gsqeu](?:,[gsqeu])*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          },
          {
            "in": "query",
            "name": "per_rail",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 60,
              "minimum": 8,
              "title": "Per Rail",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeImageRails"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Home Image Rails"
      }
    },
    "/api/home/tags": {
      "get": {
        "operationId": "home_tags_api_home_tags_get",
        "parameters": [
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:[gsqeu](?:,[gsqeu])*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          },
          {
            "in": "query",
            "name": "featured_limit",
            "required": false,
            "schema": {
              "default": 18,
              "maximum": 40,
              "minimum": 1,
              "title": "Featured Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "group_limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 80,
              "minimum": 5,
              "title": "Group Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeTags"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Home Tags"
      }
    },
...
    "/api/popularity/periods": {
      "get": {
        "operationId": "popularity_periods_api_popularity_periods_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "day",
              "pattern": "^(day|month|year)$",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder"
            }
          },
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          },
          {
            "in": "query",
            "name": "blacklist",
            "required": false,
            "schema": {
              "default": "",
              "title": "Blacklist",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PopularityPeriod"
                  },
                  "title": "Response Popularity Periods Api Popularity Periods Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Popularity Periods"
      }
    },
...
  }
}